requirements.txt
Flask==3.1.0
supabase==2.10.0
Pillow
bcrypt==4.2.1
python-dotenv==1.0.0
piexif==1.1.3
gunicorn==21.2.0
google-cloud-aiplatform==1.73.0
